feat(definition-registry): P5 — the stem-parsing and the species palette die
The npc no longer reads taxonomy off a file path. resolve_thing_in used to split
the texture stem and look the segment up in a code-owned palette — a definition's
identity derived from where its pictures live. It takes the def WHOLE from the
registry now, whose id already carries type/species/kind/variant.

That forced a shape change: the injection carries the FULL definition_reference
rather than the kind half. Dropping the type/subtype halves at bind time was
exactly what made the stem-parsing necessary.

No registry is now a hard error for a pawn, deliberately — a guessed subtype
would be adopted and rendered wrong forever. The fixture asserts both directions.

The species palette is deleted; its own comment said when it could be ('once a
content registry guarantees append-only numbering'). Species ids move to
content/subtypes.toml (F16) with animal=1 human=2 carried over, so every stored
pawn def reads the same. The asymmetry is stated: a subtype id is authored on its
own record where one exists, and in subtypes.toml where none does.

Live: edge definitions=17, npc def=0x30010070 speed 12 thirst 1, and grep finds
neither workaround anywhere.

+++ server/edge/src/worldgen.rs
@@ -75,7 +75,7 @@ impl Worldgen {
     /// corpus REORDER without renumbering — which is the fragility this whole stream removes.
     pub fn load_versioned_with(
         content_dir: &Path,
-        registry: Option<std::collections::HashMap<(bool, String), u16>>,
+        registry: Option<std::collections::HashMap<(bool, String), u32>>,
     ) -> Result<LoadedWorldgen, String> {
         let sources = resonantdust_content::content::read_content_dir(content_dir)
             .map_err(|e| format!("read content {}: {e}", content_dir.display()))?;
@@ -103,7 +103,7 @@ impl Worldgen {
     /// As [`from_sources`], with the definition registry injected when available.
     pub fn from_sources_with(
         sources: &[(String, String)],
-        registry: Option<std::collections::HashMap<(bool, String), u16>>,
+        registry: Option<std::collections::HashMap<(bool, String), u32>>,
     ) -> Result<Worldgen, String> {
         let bundle = resonantdust_content::load(sources).map_err(|errs| {
             errs.iter().map(|e| format!("{}: {}", e.file, e.message)).collect::<Vec<_>>().join("; ")
